Move decodeFullSig into decode.ts
