+++ packages/iov-bns/types/decode.d.ts
@@ -1,8 +1,9 @@
-import { Amount, ChainId, Nonce, SignedTransaction, Token, UnsignedTransaction } from "@iov/bcp";
+import { Amount, ChainId, FullSignature, Nonce, SignedTransaction, Token, UnsignedTransaction } from "@iov/bcp";
 import * as codecImpl from "./generated/codecimpl";
 import { BnsUsernameNft, CashConfiguration, Keyed, Participant } from "./types";
 export declare function decodeUsernameNft(nft: codecImpl.username.IUsernameToken, registryChainId: ChainId): BnsUsernameNft;
 export declare function decodeNonce(acct: codecImpl.sigs.IUserData & Keyed): Nonce;
+export declare function decodeFullSig(sig: codecImpl.sigs.IStdSignature): FullSignature;
 export declare function decodeToken(data: codecImpl.currency.ITokenInfo & Keyed): Token;
 export declare function decodeAmount(coin: codecImpl.coin.ICoin): Amount;
 export declare function decodeCashConfiguration(config: codecImpl.cash.Configuration): CashConfiguration;
